feat: add ability to use skip/only with test fixtures @W-15228939 (#4054)
* refactor: split jest utils into multiple files

* feat: add ability to use only/skip in test fixtures

* chore: simplify ignoring dist in .gitignore

* docs(contributing): add note on `.only` and `.skip` files

* feat: add hacky eslint rule to catch `.skip` and `.only` files

* chore: warn about .skip files when linting project, but error when trying to commit

* fix(eslintrc): wrong error level
